+++ CHANGES.rst
@@ -5,6 +5,7 @@ Wazimap Change Log
 ------------------
 
 * FIX: redirect to slugged URLs when possible
+* BREAKING: Data tables now default to one table for all geo levels
 
 0.2.2 (15 February 2016)
 ------------------------