+++ CHANGES.rst
@@ -7,6 +7,7 @@ Wazimap Change Log
 * FIX: redirect to slugged URLs when possible
 * BREAKING: Data tables now default to one table for all geo levels
 * Include deployment file examples in the ``deploy`` directory
+* Remove uservoice.js
 
 0.2.2 (15 February 2016)
 ------------------------